requirements.txt
# Core
pandas>=1.5.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0

# Data Collection
alpaca-py>=0.13.0
yfinance>=0.2.28
pyarrow>=12.0.0  # For parquet support

# NLP
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.30.0

# ML Models
scikit-learn>=1.3.0
xgboost>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utils
tqdm>=4.66.0
joblib>=1.3.0
scipy>=1.11.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0